requirements.txt
Pygame==2.3.0
PyOpenGL==3.1.6